requirements.txt
# Runway Detection System Requirements

# API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Database
psycopg2-binary==2.9.9

# HTTP Client
requests==2.31.0

# Data Processing
python-dateutil==2.8.2

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2  # For testing FastAPI

# Scheduling (for cron alternative)
schedule==1.2.0
